refactor(scripts): 收敛脚本结构并调整 CH592 电池 ADC
