+++ firmware/CH592F/hal/include/kbd_battery.h
@@ -9,6 +9,7 @@
  *
  * 硬件:
  * - PA14 (AIN4): VBAT 经 100K+100K 分压 (1/2) 后的 ADC 输入
+ * - ADC PGA: 1/4x (-12dB)
  * - PA15: VBAT_AD_EN, 高电平启动分压电路
  * - PA13: TP4054 CHRG 引脚 (开漏输出)
  *   - 低电平: 正在充电
