+++ tools/scripts/console.py
@@ -17,6 +17,7 @@ VENV_DIR = PROJECT_ROOT / ".venv"
 VENV_BIN_DIR = VENV_DIR / ("Scripts" if os.name == "nt" else "bin")
 CONSOLE_REQUIREMENTS = SCRIPT_DIR / "requirements-console.txt"
 SETUP_SCRIPT = SCRIPT_DIR / "setup.py"
+FETCH_WINDOWS_DLL_SCRIPT = PROJECT_ROOT / "tools" / "meowisp" / "scripts" / "fetch_windows_dll.py"
 
 
 def _venv_python() -> Path:
@@ -114,6 +115,40 @@ def _ensure_wchisp_ready(python_exe: Path) -> None:
         )
 
 
+def _ensure_windows_isp_assets_ready(python_exe: Path) -> None:
+    if os.name != "nt":
+        return
+    if not FETCH_WINDOWS_DLL_SCRIPT.is_file():
+        print(
+            f"[console] Windows ISP DLL fetch script not found: {FETCH_WINDOWS_DLL_SCRIPT}",
+            file=sys.stderr,
+        )
+        return
+    wchisp = find_wchisp()
+    if not wchisp:
+        print("[console] wchisp not found, skip CH375 DLL check.", file=sys.stderr)
+        return
+    dll_path = wchisp.parent / "CH375DLL64.dll"
+    if dll_path.is_file():
+        return
+
+    print(
+        f"[console] CH375DLL64.dll not found next to wchisp, fetching to {dll_path}...",
+        file=sys.stderr,
+    )
+    result = subprocess.run(
+        [str(python_exe), str(FETCH_WINDOWS_DLL_SCRIPT), "--out", str(dll_path)],
+        env=_venv_env(),
+        check=False,
+    )
+    if result.returncode != 0:
+        print(
+            "[console] Auto-fetch of CH375DLL64.dll failed. "
+            "Windows ISP may be unavailable until the DLL is downloaded.",
+            file=sys.stderr,
+        )
+
+
 def main() -> None:
     if len(sys.argv) > 1:
         raise SystemExit("Legacy console flags were removed. Run `python tools/scripts/console.py` without arguments.")
@@ -124,6 +159,7 @@ def main() -> None:
         _reexec_in_venv(python_exe)
     os.environ.update(_venv_env())
     _ensure_wchisp_ready(python_exe)
+    _ensure_windows_isp_assets_ready(python_exe)
 
     try:
         from tui_textual import run_textual
