+++ tools/scripts/targets/ch552/__init__.py
@@ -1,3 +1,4 @@
+from .build import DEFAULT_KEYBOARD, VALID_KEYBOARDS, build_dir_for
 from .profile import CH552_PROFILE
 
-__all__ = ["CH552_PROFILE"]
+__all__ = ["CH552_PROFILE", "DEFAULT_KEYBOARD", "VALID_KEYBOARDS", "build_dir_for"]
