+++ tools/scripts/targets/ch592/__init__.py
@@ -1,3 +1,4 @@
+from .build import DEFAULT_KEYBOARD, DEFAULT_PROFILE, artifact_paths, preset_for
 from .profile import CH592_PROFILE
 
-__all__ = ["CH592_PROFILE"]
+__all__ = ["CH592_PROFILE", "DEFAULT_KEYBOARD", "DEFAULT_PROFILE", "artifact_paths", "preset_for"]
