+++ tools/scripts/tui_console.py
@@ -22,7 +22,7 @@ try:
 except ImportError:  # pragma: no cover - Windows fallback
     curses = None
 
-from ch552g import VALID_KEYBOARDS as CH552_KEYBOARDS
+from targets.ch552.build import VALID_KEYBOARDS as CH552_KEYBOARDS
 from common import colorize as _c, find_cmake, find_sdcc, find_wchisp, display_path, use_color
 from i18n import t
 from targets.ch592.profile import _find_gcc_in_toolchain
